Test-db-manager: No Node.js Cypress spec bundle
Add a secondary output bundle, which is free from native Node.js API
calls and code (no knex, no pg, no utils, no fs calls, ...) and which
can safely be imported in Cypress spec/spec-helper files.

Previously (will be fixed in next commit) the full bundle was imported
from Cypress test spec files, which resulted in knex + postgresql
driver, among other, being included in the code running in the test
code running in the browser. All database calls should happen trough
Cypress tasks and run on the Node.js process side, thus there is no
need to bundle those parts into the actual test code.
