+++ test-db-manager/src/CypressSpecExports.ts
@@ -1,0 +1,24 @@
+/**
+ * Special export file that only exports types, enums and clean/simple functions
+ * that can be safely imported from within Cypress spec/helper files.
+ *
+ * All direct or indirect imports to complex node packages are forbidden,
+ * such as hsl-timetables-inserter, pg, knex. Or more specifically, no Node.js
+ * API modules can be imported.
+ *
+ * uuid is allowed to be imported as it has a proper browser API implementation,
+ * and does not need any Node API's to be polyfilled.
+ */
+
+// Include enums and types from generates GQL types and from ./types dir
+export * from './generated/graphql';
+export * from './types';
+
+// Include builder tools for datasets
+export * from './builders';
+
+// Include datasets.
+export * from './datasets';
+
+// Random file that is needed
+export * from './queries/infrastructure';
