--- test-db-manager/src/db-helpers/enums/index.ts
@@ -1,1 +1,0 @@
-export * from './hasuraEnvironment';
