+++ test-db-manager/src/db-helpers/index.ts
@@ -1,3 +1,3 @@
-export * from './enums';
 export * from './flatten';
+export * from './routesAndLines';
 export * from './timetables';
