+++ test-db-manager/src/types/enums.ts
@@ -31,3 +31,11 @@ export enum TimetablePriority {
   Draft = 30, // overrides Special, Temporary and Standard, not visible to external systems
   Staging = 40, // imported from Hastus, not in use until priority is changed
 }
+
+export enum HasuraEnvironment {
+  /** Enumeration for Hasura and database instances.
+   * This is used to route data to the same Hasura and DB instance in e2e or dev context.
+   */
+  default = 'default',
+  e2e = 'e2e',
+}
